Add TA-Lib dependency to setup.py
A lot of the TA-Lib-related problems people are having lately stems from the fact that the TA-Lib dependency is only declared in `requirements.txt`, but people just doing a `pip install cufflinks` aren't seeing that - pip doesn't care about `requirements.txt`, it only cares about `install_requires`. So this syncs it up.